// src/enums/booking.ts
export enum BookingStatus {
  PENDING_AGENT_APPROVAL = 'PENDING_AGENT_APPROVAL',
  CONFIRMED = 'CONFIRMED',
  IN_PROGRESS = 'IN_PROGRESS',
  COMPLETED = 'COMPLETED',
  CANCELLED_BY_USER = 'CANCELLED_BY_USER',
  CANCELLED_BY_AGENT = 'CANCELLED_BY_AGENT',
  CANCELLED_BY_ADMIN = 'CANCELLED_BY_ADMIN',
  REFUND_PENDING = 'REFUND_PENDING',
  REFUNDED = 'REFUNDED',
}